backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0
pydantic[email]==2.8.0
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery[redis]==5.4.0
pydicom==2.4.4
hl7apy==1.3.4
fhir.resources==7.1.0
httpx==0.27.0
reportlab==4.2.0
redis==5.0.8
python-multipart==0.0.9
aiofiles==23.2.1
psycopg2-binary==2.9.9
ruff==0.6.0
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.0
